+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v7.0-beta.5] - 2019-09-23
+
+* Fixes to proxy
+
+## [v7.0-beta.4] - 2019-09-16
+
+* Fix stroom-proxy Dockerfile
+
+## [v7.0-beta.3] - 2019-09-16
+
+* Minor fixes, including an essential fix to config
+
+## [v7.0-beta.2] - 2019-09-13
+
+* Fix docker build
+
+## [v7.0-beta.1] - 2019-09-11
+
 * Issue **#1253** : Data retention policies containing just `AND` will now match everything.
 
 * Issue **#1252** : Stream type suggestions no longer list internal types.
@@ -1725,7 +1743,12 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-alpha.5...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.5...HEAD
+[v7.0-beta.5]: https://github.com/gchq/stroom/compare/v7.0-beta.4...v7.0-beta.5
+[v7.0-beta.4]: https://github.com/gchq/stroom/compare/v7.0-beta.3...v7.0-beta.4
+[v7.0-beta.3]: https://github.com/gchq/stroom/compare/v7.0-beta.2...v7.0-beta.3
+[v7.0-beta.2]: https://github.com/gchq/stroom/compare/v7.0-beta.1...v7.0-beta.2
+[v7.0-beta.1]: https://github.com/gchq/stroom/compare/v7.0-alpha.5...v7.0-beta.1
 [v7.0-alpha.5]: https://github.com/gchq/stroom/compare/v7.0-alpha.4...v7.0-alpha.5
 [v7.0-alpha.4]: https://github.com/gchq/stroom/compare/v7.0-alpha.3...v7.0-alpha.4
 [v7.0-alpha.3]: https://github.com/gchq/stroom/compare/v7.0-alpha.2...v7.0-alpha.3